move TestResults out of components-api, remove InterpretationResult from api
