+++ engine/flink/avro-util/src/test/scala/pl/touk/nussknacker/engine/avro/source/flink/TestFromFileSpec.scala
@@ -6,7 +6,8 @@ import com.typesafe.scalalogging.LazyLogging
 import io.circe.Json._
 import org.apache.kafka.common.record.TimestampType
 import org.scalatest.{FunSuite, Matchers}
-import pl.touk.nussknacker.engine.api.deployment.TestProcess._
+import pl.touk.nussknacker.engine.testmode.TestProcess._
+import pl.touk.nussknacker.engine.api.test.TestData
 import pl.touk.nussknacker.engine.avro.KafkaAvroBaseComponentTransformer.{SchemaVersionParamName, TopicParamName}
 import pl.touk.nussknacker.engine.avro.KafkaAvroIntegrationMockSchemaRegistry.schemaRegistryMockClient
 import pl.touk.nussknacker.engine.avro.KafkaAvroTestProcessConfigCreator
