+++ engine/flink/executor/src/test/scala/pl/touk/nussknacker/engine/process/typeinformation/TypeInformationDetectionSpec.scala
@@ -8,8 +8,8 @@ import pl.touk.nussknacker.engine.api.context.ValidationContext
 import pl.touk.nussknacker.engine.api.deployment.DeploymentData
 import pl.touk.nussknacker.engine.api.typed.typing
 import pl.touk.nussknacker.engine.api.typed.typing.{Typed, Unknown}
+import pl.touk.nussknacker.engine.api.{Context, ProcessVersion, ValueWithContext}
 import pl.touk.nussknacker.engine.flink.api.typeinformation.{TypeInformationDetection, TypingResultAwareTypeInformationCustomisation}
-import pl.touk.nussknacker.engine.api.{Context, InterpretationResult, ProcessVersion, ValueWithContext}
 import pl.touk.nussknacker.engine.flink.api.{ConfigGlobalParameters, DefaultAdditionalInformationSerializer, NkGlobalParameters}
 import pl.touk.nussknacker.engine.process.typeinformation.internal.typedobject.TypedScalaMapTypeInformation
 import pl.touk.nussknacker.test.ClassLoaderWithServices
@@ -76,10 +76,6 @@ class CustomTypeInformationCustomisation extends TypingResultAwareTypeInformatio
 }
 
 class CustomTypeInformationDetection extends TypeInformationDetection {
-  override def forInterpretationResult(validationContext: ValidationContext,
-                                       output: Option[typing.TypingResult]): TypeInformation[InterpretationResult] = throw new IllegalArgumentException("Checking loader :)")
-
-  override def forInterpretationResults(results: Map[String, ValidationContext]): TypeInformation[InterpretationResult] = throw new IllegalArgumentException("Checking loader :)")
 
   override def forContext(validationContext: ValidationContext): TypeInformation[Context] = throw new IllegalArgumentException("Checking loader :)")
