+++ engine/lite/request-response/util/src/test/scala/pl/touk/nussknacker/engine/requestresponse/utils/QueryStringTestDataParserSpec.scala
@@ -3,7 +3,7 @@ package pl.touk.nussknacker.engine.requestresponse.utils
 import java.nio.charset.StandardCharsets
 import java.util
 import org.scalatest.{FunSuite, Matchers}
-import pl.touk.nussknacker.engine.api.deployment.TestProcess.TestData
+import pl.touk.nussknacker.engine.api.test.TestData
 import pl.touk.nussknacker.engine.api.typed.TypedMap
 
 class QueryStringTestDataParserSpec extends FunSuite with Matchers {
