+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/testmode/TestDataPreparer.scala
@@ -2,8 +2,8 @@ package pl.touk.nussknacker.engine.testmode
 
 import pl.touk.nussknacker.engine.ModelData
 import pl.touk.nussknacker.engine.api.NodeId
-import pl.touk.nussknacker.engine.api.deployment.TestProcess.TestData
 import pl.touk.nussknacker.engine.api.process.{ComponentUseCase, SourceTestSupport}
+import pl.touk.nussknacker.engine.api.test.TestData
 import pl.touk.nussknacker.engine.compile.ExpressionCompiler
 import pl.touk.nussknacker.engine.compile.nodecompilation.NodeCompiler
 import pl.touk.nussknacker.engine.graph.EspProcess
