+++ scenario-api/src/main/scala/pl/touk/nussknacker/engine/api/PartReference.scala
